+++ test/database/test_database.py
@@ -100,7 +100,14 @@ def test_sorted_set(dc: DataCheck):
 
 
 def test_dialect(dc: DataCheck):
-    assert dc.sql.dialect in ("sqlite", "postgresql", "mysql", "mssql", "oracle")
+    assert dc.sql.dialect in (
+        "sqlite",
+        "postgresql",
+        "mysql",
+        "mssql",
+        "oracle",
+        "duckdb",
+    )
 
 
 def test_leading_zeros_string(dc: DataCheck):